Migrate CustomType prompt
